Migrate Mutation.createOrganization
