+++ packages/services/api/src/modules/organization/validation.ts
@@ -1,0 +1,3 @@
+import { NameModel } from '../../shared/entities';
+
+export const OrganizationNameModel = NameModel.min(2).max(50);
